Minor note about overriding default values...
I added a minor note about overriding default values...once I discovered that this could be done.  (I was happy to discover this because I was not happy listing all of the values just to override one of them, for example.)

Anyway, I wasn't too happy with the wording....so someone may have a better way of putting it so that it's clear and concise like the rest of the docs.

I also broke up a couple of sentences in the section on labels.  To me, it reads better broken up....two different thoughts.

Please let me know if you have any questions/concerns.

Thx.  🤓